Remove nextjs vue/web-component examples
