+++ packages/widget/scripts/prepublish.cjs
@@ -4,8 +4,6 @@ const fs = require('fs/promises');
 const path = require('path');
 const packageJson = require('../package.json');
 const coreVersion = require('@skip-go/client/package.json').version;
-const webComponentPackageJson = require('./create-web-component-package-json.cjs');
-
 
 async function prepublish() {
   delete packageJson.scripts;
@@ -15,17 +13,6 @@ async function prepublish() {
   await fs.writeFile(targetPath, JSON.stringify(packageJson, null, 2), {
     encoding: 'utf-8',
   });
-
- // Prepare web component package
- const webComponentDir = path.resolve(__dirname, '../widget-web-component');
- await fs.mkdir(webComponentDir, { recursive: true });
- 
- // Write web component package.json
- await fs.writeFile(
-   path.resolve(webComponentDir, 'package.json'),
-   JSON.stringify(webComponentPackageJson(), null, 2),
-   { encoding: 'utf-8' }
- );
 }
 
 void prepublish();
