+++ packages/widget/widget-web-component/package.json
@@ -2,10 +2,7 @@
   "name": "@skip-go/widget-web-component",
   "description": "Swap widget web component",
   "version": "2.4.3",
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/skip-mev/widget.git"
-  },
+  "repository": "https://github.com/skip-mev/widget",
   "type": "module",
   "exports": {
     ".": {
@@ -16,18 +13,42 @@
   "engines": {
     "node": ">=12.4"
   },
+  "files": [
+    "build",
+    "README.md"
+  ],
   "scripts": {
     "publish": "npm publish widget-web-component --access public"
   },
-  "files": [
-    "dist",
-    "README.md"
-  ],
   "peerDependencies": {
     "react": "17.x || 18.x",
     "react-dom": "17.x || 18.x",
     "styled-components": "^6.0.0",
     "stylis": "^4.0.0"
+  },
+  "devDependencies": {
+    "@rollup/plugin-url": "^8.0.2",
+    "@types/node": "18.7.18",
+    "@types/qrcode": "^1.4.2",
+    "@types/react": "^18.0.6",
+    "@types/react-dom": "^18.0.2",
+    "autoprefixer": "^10.4.19",
+    "node-polyfill-webpack-plugin": "^4.0.0",
+    "postcss": "^8.4.38",
+    "postcss-loader": "^8.1.1",
+    "postcss-prefixwrap": "^1.49.0",
+    "raw-loader": "^4.0.2",
+    "rollup-plugin-copy": "^3.5.0",
+    "styled-components": "^6.0.0",
+    "stylis": "^4.0.0",
+    "tailwindcss": "^3.4.3",
+    "tailwindcss-animate": "^1.0.7",
+    "ts-loader": "^9.5.1",
+    "typed-query-selector": "^2.11.2",
+    "typescript": "5.2.x",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4"
   },
   "resolutions": {
     "@cosmos-kit/cosmostation-extension": "^2.7.10",
@@ -67,7 +88,7 @@
     "@radix-ui/react-scroll-area": "^1.0.5",
     "@radix-ui/react-switch": "^1.0.3",
     "@radix-ui/react-tooltip": "^1.0.7",
-    "@skip-go/client": "0.7.0",
+    "@skip-go/client": "workspace:packages/client",
     "@solana/spl-token": "^0.4.6",
     "@solana/wallet-adapter-react": "^0.15.35",
     "@solana/wallet-adapter-wallets": "^0.19.32",
@@ -87,19 +108,5 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "main": "index.js",
-  "keywords": [
-    "@skip-go",
-    "skip-go",
-    "widget",
-    "web-component",
-    "ibc-fun"
-  ],
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/skip-mev/widget/issues"
-  },
-  "homepage": "https://github.com/skip-mev/widget#readme"
+  }
 }
